pyproject.toml
[project]
name = "dhub"
version = "0.1.0"
description = "DataHub CLI application for database operations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faker>=22.0.0",
    "psycopg[binary]>=3.2.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "shellingham>=1.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dhub = "dhub.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []
